+++ CHANGELOG.md
@@ -5,9 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 =======
 
 ## UNRELEASED
+
+### **Added**
+
+### **Changed**
+- modified `fsx-lustre` module to auto-import data if configured
+
+### **Removed**
+
+
+=======
+
+
+
+=======
+
+## v1.4.0 (2024-02-27)
 
 ### **Added**